refactor: studyservice: splitted into 3 separate: For read, create/update and delete
also splitted the test classes for the service
