+++ src/Sepes.Infrastructure/Dto/Sandbox/SandboxListItemDto.cs
@@ -1,0 +1,8 @@
+﻿namespace Sepes.Infrastructure.Dto.Sandbox
+{
+    public class SandboxListItemDto
+    {
+        public int Id { get; set; }
+        public string Name { get; set; }
+    }
+}
