--- src/Sepes.Infrastructure/Service/Interface/IStudyService.cs
@@ -1,29 +1,0 @@
-﻿using Sepes.Infrastructure.Constants;
-using Sepes.Infrastructure.Dto.Study;
-using System.Collections.Generic;
-using System.Threading.Tasks;
-
-namespace Sepes.Infrastructure.Service.Interface
-{
-    public interface IStudyService
-    {        
-        Task<IEnumerable<StudyListItemDto>> GetStudyListAsync(bool? excludeHidden = null);
-        Task<StudyDto> GetStudyDtoByIdAsync(int studyId, UserOperation userOperation);
-
-        Task<StudyDetailsDto> GetStudyDetailsDtoByIdAsync(int studyId, UserOperation userOperation);
-
-        Task<StudyDetailsDto> CreateStudyAsync(StudyCreateDto newStudy);
-
-        Task<StudyDetailsDto> UpdateStudyMetadataAsync(int studyId, StudyDto newStudy);
-
-        Task<StudyResultsAndLearningsDto> GetResultsAndLearningsAsync(int studyId);
-        Task<StudyResultsAndLearningsDto> UpdateResultsAndLearningsAsync(int studyId, StudyResultsAndLearningsDto resultsAndLearnings);
-
-       
-
-        Task CloseStudyAsync(int studyId);
-        Task DeleteStudyAsync(int studyId);
-
-   
-    }
-}
