+++ src/Sepes.Infrastructure/Service/ServiceBase.cs
@@ -64,6 +64,5 @@ namespace Sepes.Infrastructure.Service
         {
             return await StudySingularQueries.GetStudyByIdCheckAccessOrThrow(_db, _userService, studyId, userOperation, withIncludes);
         }
-
     }
 }
