+++ src/Sepes.Infrastructure/Service/Study/Interface/IStudyDeleteService.cs
@@ -1,0 +1,10 @@
+﻿using System.Threading.Tasks;
+
+namespace Sepes.Infrastructure.Service.Interface
+{
+    public interface IStudyDeleteService
+    {  
+        Task CloseStudyAsync(int studyId);
+        Task DeleteStudyAsync(int studyId);   
+    }
+}
