+++ src/Sepes.Infrastructure/Service/Study/Interface/IStudyLogoService.cs
