+++ src/Sepes.Infrastructure/Service/Study/StudyLogoService.cs
