+++ src/Sepes.Tests/Services/DomainServices/Study/StudyServiceTestBase.cs
@@ -1,0 +1,10 @@
+﻿namespace Sepes.Tests.Services.DomainServices
+{
+    public class StudyServiceTestBase : ServiceTestBase
+    {
+        public StudyServiceTestBase()
+            :base()
+        {
+        }
+    }
+}
