+++ src/Sepes.Tests/Services/DomainServices/Study/StudyServiceTests.cs
@@ -1,0 +1,28 @@
+using Sepes.Infrastructure.Constants;
+using Sepes.Tests.Setup;
+using Xunit;
+
+namespace Sepes.Tests.Services.DomainServices
+{
+    public class StudyServiceTests : StudyServiceTestBase
+    {  
+        public StudyServiceTests()
+            :base()
+        {
+          
+        }
+        
+        [Theory]
+        [InlineData(2)]
+        [InlineData(7)]
+        [InlineData(99999)]
+        [InlineData(123456)]
+        public async void GetStudyByIdAsync_WillThrow_IfStudyDoesNotExist(int id)
+        {
+            await ClearTestDatabase();
+            var studyService = StudyServiceMockFactory.Create(_serviceProvider);
+
+            await Assert.ThrowsAsync<Infrastructure.Exceptions.NotFoundException>(() => studyService.GetStudyDtoByIdAsync(id, UserOperation.Study_Read));
+        }       
+    }
+}
