+++ src/Sepes.Tests/Setup/StudyServiceMockFactory.cs
@@ -17,11 +17,36 @@ namespace Sepes.Tests.Setup
             var mapper = serviceProvider.GetService<IMapper>();
             var logger = serviceProvider.GetService<ILogger<StudyService>>();
             var userService = UserFactory.GetUserServiceMockForAdmin(1);
+
+            var logoServiceMock = new Mock<IStudyLogoService>();
+
+            return new StudyService(db, mapper, logger, userService.Object, logoServiceMock.Object);
+        }
+
+        public static IStudyCreateUpdateService CreateUpdateService(ServiceProvider serviceProvider)
+        {
+            var db = serviceProvider.GetService<SepesDbContext>();
+            var mapper = serviceProvider.GetService<IMapper>();
+            var logger = serviceProvider.GetService<ILogger<StudyService>>();
+            var userService = UserFactory.GetUserServiceMockForAdmin(1);
             var studySpecificDatasetService = DatasetServiceMockFactory.GetStudySpecificDatasetService(serviceProvider);
 
             var logoServiceMock = new Mock<IStudyLogoService>();
 
-            return new StudyService(db, mapper, logger, userService.Object, logoServiceMock.Object, studySpecificDatasetService);
-        }      
+            return new StudyCreateUpdateService(db, mapper, logger, userService.Object, logoServiceMock.Object, studySpecificDatasetService);
+        }
+
+        public static IStudyDeleteService DeleteService(ServiceProvider serviceProvider)
+        {
+            var db = serviceProvider.GetService<SepesDbContext>();
+            var mapper = serviceProvider.GetService<IMapper>();
+            var logger = serviceProvider.GetService<ILogger<StudyService>>();
+            var userService = UserFactory.GetUserServiceMockForAdmin(1);
+            var studySpecificDatasetService = DatasetServiceMockFactory.GetStudySpecificDatasetService(serviceProvider);
+
+            var logoServiceMock = new Mock<IStudyLogoService>();
+
+            return new StudyDeleteService(db, mapper, logger, userService.Object, logoServiceMock.Object, studySpecificDatasetService);
+        }
     }
 }
